~ updated to AFNetworking 2.0 ~ added privacy policy and terms of service ~ now explicitly removing residence on sign out

+++ Classes/VLBScrollView.m
@@ -256,7 +256,7 @@ return view;
     UITapGestureRecognizer *tapGestureRecognizer = (UITapGestureRecognizer*)sender;
     CGPoint locationInView = [tapGestureRecognizer locationInView:self];
     NSUInteger index = [self indexOf:locationInView];
-    DDLogVerbose(@"%luu", index);
+    DDLogVerbose(@"%luu", (unsigned long)index);
     
     NSUInteger numberOfViews = [self.datasource numberOfViewsInScrollView:self];
